requirements.txt
asgiref==3.8.1
Django==5.1.2
django-phonenumber-field==8.0.0
gunicorn==23.0.0
packaging==25.0
phonenumbers==8.13.48
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
